requirements.txt
jellyfin-apiclient-python
python-dotenv